sync: Stop using SyncOpBase::Record for render pass
